Abstract insert new user to firestore
Next step unit testing #38
