+++ Sweat Social/FirebaseAPI/FirebaseFirestoreService.swift
@@ -1,0 +1,31 @@
+//
+//  FirebaseFirestoreService.swift
+//  Sweat Social
+//
+//  Created by Jack.Knox on 2024-03-04.
+//
+
+import Foundation
+import FirebaseFirestore
+
+class FirebaseFirestoreService : FirestoreProtocol {
+    
+    func insertNewUser(id: String, name: String, email: String, completion: @escaping (Result<Void?, Error>) -> Void) {
+        let newUser = User(id: id,
+                           name: name,
+                           email: email,
+                           joined: Date().timeIntervalSince1970) // Firebase cannot store date as is, so this is a way to handle that
+        
+        let db = Firestore.firestore()
+        
+        db.collection("users")
+            .document(id)
+            .setData(newUser.asDictionary()) { error in
+                if let error = error {
+                    completion(.failure(error))
+                } else {
+                    completion(.success(()))
+                }
+            }
+    }
+}
